feat(docs): migrate shared-layout-animation to ExampleV2
